Initial commit
This initial commit is almost exactly what Django 1.11.5 generates when
the following command is issued:

    django-admin startproject mysite

There was one alteration made; the SECRET_KEY Django populates in settings.py
is best kept secret. As such it shoudln't be checked in to a public repo. This
can instead be source from an environement variable.