Updated stylesheet
More specifically, image and icon stylesheet updated. Images will now
have a translucent overlay applied to them, as well as an embedded icon
(if specified).

Once hovered, the overlay opacity will increase. The icon will also grow
in size. This icon is a hint as to the topic of the work item.