Created base template
A base template has been created which will act as the skeleton for
all pages with in application. It ensures each page conforms to a
given style. It also provides a common header and footer for all pages.

Varying attributes in the base template (e.g. banner text and image)
have been made modifiable.

Urls namespaced.